Split off code required by libbitcoinconsensus
