+++ src/Makefile.am
@@ -103,12 +103,23 @@ endif
 LIBMW_CPPFLAGS = -I$(srcdir) -I$(srcdir)/libmw/include -I$(srcdir)/libmw/deps/caches/include -I$(srcdir)/libmw/deps/crypto/include -I$(srcdir)/libmw/deps/ghc/include  -I$(srcdir)/libmw/deps/mio/include
 libmw_libmw_a_CPPFLAGS = $(AM_CPPFLAGS) $(BITCOIN_INCLUDES) $(LIBMW_CPPFLAGS)
 libmw_libmw_a_CXXFLAGS = $(AM_CXXFLAGS) $(PIE_FLAGS)
+
+libmw_consensus_SOURCES = \
+  libmw/src/consensus/libmw.cpp \
+  libmw/src/consensus/Commitment.cpp \
+  libmw/src/consensus/Kernel.cpp \
+  libmw/src/consensus/Output.cpp \
+  libmw/src/consensus/Transaction.cpp \
+  libmw/src/consensus/TxBody.cpp \
+  libmw/src/crypto/Hasher.cpp \
+  libmw/src/traits/Serializable.cpp
+
 libmw_libmw_a_SOURCES = \
+	$(libmw_consensus_SOURCES) \
 	libmw/src/common/Logger.cpp \
 	libmw/src/crypto/Bulletproofs.cpp \
 	libmw/src/crypto/ConversionUtil.cpp \
 	libmw/src/crypto/Crypto.cpp \
-	libmw/src/crypto/Hasher.cpp \
 	libmw/src/crypto/MuSig.cpp \
 	libmw/src/crypto/Pedersen.cpp \
 	libmw/src/crypto/PublicKeys.cpp \
@@ -148,7 +159,6 @@ libmw_libmw_a_SOURCES = \
 	libmw/src/node/ICoinsView.cpp \
 	libmw/src/node/Node.cpp \
 	libmw/src/node/Snapshot.cpp \
-	libmw/src/traits/Serializable.cpp \
 	libmw/src/wallet/Keychain.cpp \
 	libmw/src/wallet/Transact.cpp \
 	libmw/src/wallet/Wallet.cpp
@@ -654,14 +664,18 @@ litecoin_wallet_LDADD += $(BOOST_LIBS) $(BDB_LIBS) $(CRYPTO_LIBS) $(MINIUPNPC_LI
 # bitcoinconsensus library #
 if BUILD_BITCOIN_LIBS
 include_HEADERS = script/bitcoinconsensus.h
-libbitcoinconsensus_la_SOURCES = $(crypto_libbitcoin_crypto_base_a_SOURCES) $(libbitcoin_consensus_a_SOURCES)
+libbitcoinconsensus_la_SOURCES = \
+  $(crypto_libbitcoin_crypto_base_a_SOURCES) \
+  $(libbitcoin_consensus_a_SOURCES) \
+  $(libmw_consensus_SOURCES) \
+  support/cleanse.cpp
 
 if GLIBC_BACK_COMPAT
   libbitcoinconsensus_la_SOURCES += compat/glibc_compat.cpp
 endif
 
 libbitcoinconsensus_la_LDFLAGS = $(AM_LDFLAGS) -no-undefined $(RELDFLAGS)
-libbitcoinconsensus_la_LIBADD = $(LIBSECP256K1) $(CRYPTO_LIBS)
+libbitcoinconsensus_la_LIBADD = $(LIBSECP256K1) $(CRYPTO_LIBS) $(BOOST_SYSTEM_LIB) $(MW_LIBS)
 libbitcoinconsensus_la_CPPFLAGS = $(AM_CPPFLAGS) -I$(builddir)/obj -I$(srcdir)/secp256k1-zkp/include -DBUILD_BITCOIN_INTERNAL $(SSL_CFLAGS) $(LIBMW_CPPFLAGS)
 libbitcoinconsensus_la_CXXFLAGS = $(AM_CXXFLAGS) $(PIE_FLAGS)
