+++ src/libmw/include/mw/models/tx/KernelBits.h
@@ -1,0 +1,9 @@
+#pragma once
+
+#include <cstdint>
+
+static const uint8_t FEE_FEATURE_BIT = 0x01;
+static const uint8_t PEGIN_FEATURE_BIT = 0x02;
+static const uint8_t PEGOUT_FEATURE_BIT = 0x04;
+static const uint8_t HEIGHT_LOCK_FEATURE_BIT = 0x08;
+static const uint8_t EXTRA_DATA_FEATURE_BIT = 0x10;
